Refactor app structure: replace ContentView with MainView, add view models for Home, AllCards, and Settings
