+++ NetPriorityAI/Models/Item.swift
@@ -2,7 +2,7 @@
 //  Item.swift
 //  NetPriorityAI
 //
-//  Created by Howard Private on 9/8/25.
+//  Created by Lung Hao Tung on 9/8/25.
 //
 
 import Foundation
