+++ NetPriorityAI/ViewModels/AllCardsViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  AllCardsViewModel.swift
+//  NetPriorityAI
+//
+//  Created by Lung Hao Tung on 9/13/25.
+//
+
+import Foundation
+
+@MainActor
+class AllCardsViewModel: ObservableObject {
+    // Future implementation for all cards functionality
+}
