+++ NetPriorityAI/ViewModels/SettingsViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  SettingsViewModel.swift
+//  NetPriorityAI
+//
+//  Created by Lung Hao Tung on 9/13/25.
+//
+
+import Foundation
+
+@MainActor
+class SettingsViewModel: ObservableObject {
+    // Future implementation for settings functionality
+}
